ApiHandler / ChainData refactoring
Removed a lot of retry/reconnect logic from ApiHandler and ChainData, as
it was spawning a lot of PJS instances, which also handling reconnects
on its own.
This change also properly adpots the idea of talking both to relay chain
and to parachains, depending on queried data.

With new logic, each ApiHandler instance governs an interface to a
single chain, where it can switch endpoints and do reconnects on its
own.
`ApiHandler.getApi()` returns a PJS instance that should work, and will
block in case we're reconnecting.

PeopleApi endpoints now live in configs, instead of being hardcoded.

Instead of passing around ApiHandlers, passing a ChainData instance now, as
it contains both ApiHandlers instance now, and as it is an intended
abstraction for it